+++ c4e-persistence-core/src/main/java/com/com4energy/persistence/clientes/BaseClienteRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.repository.NoRepositoryBean;
 
 @NoRepositoryBean
-public interface BaseClienteRepository extends JpaRepository<Cliente, Long> {
+public interface BaseClienteRepository extends JpaRepository<Cliente, Integer> {
 
 }
